Refactor monolithic app to a modular structure
